requirements.txt
pipecat-ai[daily,openai,deepgram,cartesia]
fastapi
uvicorn
python-dotenv
loguru
aiohttp
xmltodict 
pipecat-ai-flows
usaddress==0.5.10